requirements.txt
fastapi-cache2==0.1.3.4
grpc-requests==0.0.10
grpcio-tools==1.37.1
homi==0.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.2.0
python-multipart==0.0.5
requests==2.25.1
